NWRTranscription Update for Checking/Correction
The procedures originally in NWRTranscription task were removed, so that
these procedures could be used by the checking and correction scripts.

The addition of or/ar as Other vowel categories, as well as ? for Other
consonant category.

The script was adjusted so that non-English WorldBet transcriptions
would be displayed and scored correctly.